move TypedProperties to basic package
