+++ sharding-orchestration/sharding-orchestration-center/sharding-orchestration-center-apollo/src/main/java/org/apache/shardingsphere/orchestration/center/instance/ApolloPropertiesEnum.java
@@ -21,7 +21,7 @@ import com.ctrip.framework.apollo.core.ConfigConsts;
 import com.ctrip.framework.apollo.openapi.client.constant.ApolloOpenApiConstants;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
-import org.apache.shardingsphere.underlying.common.properties.TypedPropertiesKey;
+import org.apache.shardingsphere.underlying.common.properties.basic.TypedPropertiesKey;
 
 /**
  * Apollo properties enum.
