+++ sharding-orchestration/sharding-orchestration-center/sharding-orchestration-center-nacos/src/main/java/org/apache/shardingsphere/orchestration/center/instance/NacosPropertiesEnum.java
@@ -19,7 +19,7 @@ package org.apache.shardingsphere.orchestration.center.instance;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
-import org.apache.shardingsphere.underlying.common.properties.TypedPropertiesKey;
+import org.apache.shardingsphere.underlying.common.properties.basic.TypedPropertiesKey;
 
 /**
  * Nacos properties enum.
