+++ sharding-orchestration/sharding-orchestration-center/sharding-orchestration-center-zookeeper-curator/src/main/java/org/apache/shardingsphere/orchestration/center/instance/ZookeeperProperties.java
@@ -17,7 +17,7 @@
 
 package org.apache.shardingsphere.orchestration.center.instance;
 
-import org.apache.shardingsphere.underlying.common.properties.TypedProperties;
+import org.apache.shardingsphere.underlying.common.properties.basic.TypedProperties;
 
 import java.util.Properties;
